move p2p tests
